# packages/api-types/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MarkRead": {
        "additionalProperties": false,
        "properties": {
          "all": {
            "default": false,
            "title": "All",
            "type": "boolean"
          },
          "ids": {
            "anyOf": [
              {
                "items": {
                  "format": "uuid",
                  "type": "string"
                },
                "maxItems": 100,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ids"
          }
        },
        "title": "MarkRead",
        "type": "object"
      },
...
      "MediaUrls": {
        "properties": {
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "variants": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Variants",
            "type": "object"
          }
        },
        "required": [
          "variants",
          "expires_at"
        ],
        "title": "MediaUrls",
        "type": "object"
      },
...
      "NotificationActor": {
        "properties": {
          "account_type": {
            "enum": [
              "private",
              "business"
            ],
            "title": "Account Type",
            "type": "string"
          },
          "can_open": {
            "title": "Can Open",
            "type": "boolean"
          },
          "nickname": {
            "title": "Nickname",
            "type": "string"
          }
        },
        "required": [
          "nickname",
          "account_type",
          "can_open"
        ],
        "title": "NotificationActor",
        "type": "object"
      },
      "NotificationOut": {
        "properties": {
          "actor": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NotificationActor"
              },
              {
                "type": "null"
              }
            ]
          },
          "body": {
            "title": "Body",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "post": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NotificationPost"
              },
              {
                "type": "null"
              }
            ]
          },
          "read": {
            "title": "Read",
            "type": "boolean"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "enum": [
              "follow_request",
              "new_follower",
              "follow_accepted",
              "vote_milestone",
              "moderation",
              "appeal_decided"
            ],
            "title": "Type",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "title",
          "body",
          "url",
          "created_at",
          "read",
          "actor",
          "post"
        ],
        "title": "NotificationOut",
        "type": "object"
      },
      "NotificationPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NotificationOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "unread": {
            "title": "Unread",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "next_cursor",
          "unread"
        ],
        "title": "NotificationPage",
        "type": "object"
      },
      "NotificationPost": {
        "properties": {
          "blurhash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blurhash"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "thumb": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MediaUrls"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "blurhash",
          "thumb"
        ],
        "title": "NotificationPost",
        "type": "object"
      },
      "NotificationSettings": {
        "properties": {
          "follows": {
            "title": "Follows",
            "type": "boolean"
          },
          "moderation": {
            "title": "Moderation",
            "type": "boolean"
          },
          "votes": {
            "title": "Votes",
            "type": "boolean"
          }
        },
        "required": [
          "follows",
          "votes",
          "moderation"
        ],
        "title": "NotificationSettings",
        "type": "object"
      },
      "NotificationSettingsPatch": {
        "additionalProperties": false,
        "properties": {
          "follows": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Follows"
          },
          "moderation": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Moderation"
          },
          "votes": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Votes"
          }
        },
        "title": "NotificationSettingsPatch",
        "type": "object"
      },
...
      "PushTokenIn": {
        "additionalProperties": false,
        "properties": {
          "platform": {
            "enum": [
              "ios",
              "android"
            ],
            "title": "Platform",
            "type": "string"
          },
          "token": {
            "maxLength": 250,
            "minLength": 20,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token",
          "platform"
        ],
        "title": "PushTokenIn",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "WearX API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/me/notification-settings": {
      "get": {
        "operationId": "get_settings__v1_me_notification_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationSettings"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Settings ",
        "tags": [
          "notifications"
        ]
      },
      "patch": {
        "description": "Quali push ricevere. La lista nell'app resta completa.",
        "operationId": "patch_settings_v1_me_notification_settings_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationSettingsPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationSettings"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Settings",
        "tags": [
          "notifications"
        ]
      }
    },
...
    "/v1/me/push-tokens": {
      "put": {
        "description": "Registra il telefono per i push. Lo stesso telefono passato a un altro account cambia\nproprietario (un telefono riceve i push di un solo account: quello con cui è entrato).",
        "operationId": "register_token_v1_me_push_tokens_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushTokenIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Token",
        "tags": [
          "notifications"
        ]
      }
    },
    "/v1/me/push-tokens/{token}": {
      "delete": {
        "description": "All'uscita dall'account: quel telefono non riceve più push. Ripetibile.",
        "operationId": "remove_token_v1_me_push_tokens__token__delete",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Token",
        "tags": [
          "notifications"
        ]
      }
    },
...
    "/v1/notifications": {
      "get": {
        "operationId": "notifications_v1_notifications_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Notifications",
        "tags": [
          "notifications"
        ]
      }
    },
    "/v1/notifications/read": {
      "post": {
        "operationId": "mark_read_v1_notifications_read_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkRead"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mark Read",
        "tags": [
          "notifications"
        ]
      }
    },
...
  }
}